+++ components/openapi/src/main/scala/pl/touk/nussknacker/openapi/parser/SecuritiesParser.scala
@@ -4,18 +4,22 @@ import cats.data.{NonEmptyList, Validated}
 import cats.implicits._
 import com.typesafe.scalalogging.LazyLogging
 import io.swagger.v3.oas.models.security.{SecurityRequirement, SecurityScheme}
+import io.swagger.v3.oas.models.security.SecurityScheme.Type.{APIKEY, HTTP}
 import pl.touk.nussknacker.engine.api.util.ReflectUtils
 import pl.touk.nussknacker.openapi.{
   ApiKeyInCookie,
   ApiKeyInHeader,
   ApiKeyInQuery,
   ApiKeySecret,
+  BasicAuth,
+  HttpBasicAuthSecret,
   Secret,
   SecurityConfig,
   SecuritySchemeName,
   SwaggerSecurity
 }
 import pl.touk.nussknacker.openapi.parser.ParseToSwaggerService.ValidationResult
+import pl.touk.nussknacker.openapi.parser.SecuritiesParser.getApiKeySecurity
 
 import scala.jdk.CollectionConverters._
 
@@ -37,7 +41,7 @@ private[parser] object SecuritiesParser extends LazyLogging {
             // finds the first security requirement that can be met by the config
             securityRequirementsDefinition.view
               .map { securityRequirement =>
-                matchSecuritiesForRequiredSchemes(
+                matchSecretsForRequiredSchemes(
                   securityRequirement.asScala.keys.toList,
                   securitySchemes,
                   securityConfig
@@ -50,42 +54,45 @@ private[parser] object SecuritiesParser extends LazyLogging {
         }
     }
 
-  private def matchSecuritiesForRequiredSchemes(
+  private def matchSecretsForRequiredSchemes(
       requiredSchemesNames: List[String],
       securitySchemes: Map[String, SecurityScheme],
       securitiesConfig: SecurityConfig
   ): ValidationResult[List[SwaggerSecurity]] =
     requiredSchemesNames.map { schemeName =>
       {
-        val validatedSecurityScheme: ValidationResult[SecurityScheme] = Validated.fromOption(
-          securitySchemes.get(schemeName),
-          NonEmptyList.of(s"""there is no security scheme definition for scheme name "$schemeName"""")
-        )
-        val validatedSecuritySecretConfigured: ValidationResult[Secret] = Validated.fromOption(
-          securitiesConfig.secret(SecuritySchemeName(schemeName)),
-          NonEmptyList.of(s"""there is no security secret configured for scheme name "$schemeName"""")
-        )
-
-        (validatedSecurityScheme, validatedSecuritySecretConfigured)
-          .mapN { case (securityScheme, configuredSecret) =>
-            getSecurityFromSchemeAndSecret(securityScheme, configuredSecret)
-          }
-          .andThen(identity)
+        val validatedSecurityScheme: ValidationResult[SecurityScheme] = Validated
+          .fromOption(
+            securitySchemes.get(schemeName),
+            NonEmptyList.of(s"""there is no security scheme definition for scheme name "$schemeName"""")
+          )
+        validatedSecurityScheme
+          .andThen { scheme => matchSecretForScheme(scheme, SecuritySchemeName(schemeName), securitiesConfig) }
       }
     }.sequence
 
-  private def getSecurityFromSchemeAndSecret(
-      securityScheme: SecurityScheme,
-      secret: Secret
-  ): ValidationResult[SwaggerSecurity] = {
-    import SecurityScheme.Type._
-    (securityScheme.getType, secret) match {
-      case (APIKEY, apiKeySecret: ApiKeySecret) =>
-        getApiKeySecurity(securityScheme, apiKeySecret).validNel
-      case (otherType: SecurityScheme.Type, _) => {
-        val secretClassName = ReflectUtils.simpleNameWithoutSuffix(secret.getClass)
-        s"Security type $otherType is not supported yet or ($otherType, $secretClassName) is a mismatch security scheme type and security config pair".invalidNel
+  private def matchSecretForScheme(
+      scheme: SecurityScheme,
+      schemeName: SecuritySchemeName,
+      securitiesConfig: SecurityConfig
+  ) = {
+    def securitySchemeNotFoundError: String =
+      s"""there is no security config for scheme name "${schemeName.value}""""
+
+    (scheme.getType, scheme.getScheme) match {
+      case (APIKEY, _) => {
+        securitiesConfig.apiKeySecret(schemeName) match {
+          case Some(secret) => getApiKeySecurity(scheme, secret).validNel
+          case None         => securitySchemeNotFoundError.invalidNel
+        }
       }
+      case (HTTP, "basic") => {
+        securitiesConfig.httpBasicAuthSecret(schemeName) match {
+          case Some(secret) => BasicAuth(schemeName.value, secret.username, secret.password).validNel
+          case None         => securitySchemeNotFoundError.invalidNel
+        }
+      }
+      case (otherType: SecurityScheme.Type, _) => s"Security type $otherType is not supported".invalidNel
     }
   }